fixed offset fields

+++ sear/irrseq00/irrseq00.hpp
@@ -254,7 +254,7 @@ typedef struct {
 /*************************************************************************/
 
 typedef struct {
-  uint32_t length;
+  uint16_t length;
   char data;
 } offset_field_t;
